refactor: remove reference to old from core
