+++ falcon/src/core/consts.ts
@@ -1,0 +1,2 @@
+export const HIST_TYPE = Int32Array;
+export const CUM_ARR_TYPE = Float32Array; // use float because we are storing sums
