+++ falcon/src/core/views/view0D.ts
@@ -1,5 +1,5 @@
 import { ViewAbstract } from "./viewAbstract";
-import { Interval } from "../../old/basic";
+import { Interval } from "../util";
 import type { Falcon } from "../falcon";
 
 /* defines how the parameter is typed for on change */
